requirements.txt
playwright
fastapi
uvicorn[standard]
python-multipart
boto3
requests
google-genai
